Fetch products on Cart page reload
When reloading the cart page the products disappear in the redux state. So changing the quantity doesn't work anymore, since it can't match products in the redux state anymore. It gives the error getState().products is null.

Also, if you used a filter on the product page, then only the filtered products would be in the redux state. If you added already products before you filtered, you also can't change the quantitiy of those.

Fetching the products fixes that issue